requirements.txt
streamlit==1.35.0
openai==1.30.3
python-dotenv==1.0.1
langchain==0.2.1
langchain-community==0.2.1
flask==3.0.3